requirements.txt
numpy==1.21.5
Pillow==9.2.0
tensorflow==2.9.1
